% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{add_csv_stub}
\alias{add_csv_stub}
\title{Helper function to check if list of countries needs .csv added at the end or not}
\usage{
add_csv_stub(x)
}
\arguments{
\item{x}{object to which you want to add .csv to the end of}
}
\description{
Helper function to check if list of countries needs .csv added at the end or not
}
\examples{
\dontrun{
add_csv_stub(c("Zambia", "Zambia.csv"))
}
}
\keyword{internal}
