% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_google_mobility.R
\name{download_google_mobility}
\alias{download_google_mobility}
\title{Download the Google mobility data for the US or a selected US state.}
\usage{
download_google_mobility(type = "states", silent = FALSE)
}
\arguments{
\item{type}{The type of geographic region for which data is to be returned.
This is one of 'US', 'states', 'counties', or a US state name.}

\item{silent}{Default is FALSE. Indicates whether a message describing the download
should be displayed.}
}
\value{
A tibble containing the Google community mobility reports data
}
\description{
Download the Google mobility data for the US or a selected US state.
}
\examples{
gmcr <- download_google_mobility(type = "Washington", silent = FALSE)
}
