% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{first_time_flag}
\alias{first_time_flag}
\title{Flags the first time and event occurs}
\usage{
first_time_flag(var, value)
}
\arguments{
\item{var}{column name on which you wish to run funciton}

\item{value}{value for which you are searching for first occurence}
}
\description{
Returns an indicator variable for the first time a condition is met in a
dataframe. Useful for tagging when an event started or ended in a panel dataset.
}
\keyword{internal}
