% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vax_countries.R
\name{get_vax_countries}
\alias{get_vax_countries}
\title{Pull list of countries for which vax data exist}
\usage{
get_vax_countries(url = owid_vax_url)
}
\arguments{
\item{url}{what is the url for the github repo containing vax info by country}
}
\description{
Returns an object of the countries being tracked by Our World in Data
}
\details{
Import Our World in Data Vaccine rates
}
\examples{
\dontrun{
vax_list <- get_vax_countries(url)
}
}
