% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_links.R
\docType{data}
\name{hdx_govmes_linkid}
\alias{hdx_govmes_linkid}
\title{HDX COVID-19 government measures linkid for parsing}
\format{
An object of class \code{character} of length 1.
}
\usage{
hdx_govmes_linkid
}
\description{
Excel file of government measures taken by diffferent countries
}
\examples{
\dontrun{
glamr::extract_excel_data(hdx_govmes_url, hdx_govmes_url, 2, 'xlsx') }
}
\keyword{datasets}
