% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_links.R
\docType{data}
\name{hdx_govmes_url}
\alias{hdx_govmes_url}
\title{HDX COVID-19 government measures}
\format{
An object of class \code{character} of length 1.
}
\usage{
hdx_govmes_url
}
\description{
Excel file of government measures taken by diffferent countries
}
\keyword{datasets}
