% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ihme_download.R
\name{ihme_download}
\alias{ihme_download}
\title{Download the IHME projection from a specified projection date}
\usage{
ihme_download(IHMEdate)
}
\arguments{
\item{IHMEdate}{One of the dates on which IHME posted a projection.
Valid dates are March 26, March 27, March 29, March 30, March 31, April 1,
April 5, April 7, April 8, April 10, April 13, April 17, April 21, April 22,
April 27, April 28, April 29, May 4, and May 10}
}
\value{
A tibble containing the IHME projection
}
\description{
Download the IHME projection from a specified projection date
}
\examples{
ihme_projection <- ihme_download("April 22")
}
