% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ihme_extract_US.R
\name{ihme_extractUS}
\alias{ihme_extractUS}
\title{Extract the data for the 50 US states and the District of Columbia from a dataframe
containing an IHME projection.}
\usage{
ihme_extractUS(dat)
}
\arguments{
\item{dat}{A dataframe or tibble with an IHME projection.}
}
\value{
A tibble with the projection for just the 50 US states and the District
of Columbia. The column naming the geographic area is renamed to \code{state}.
}
\description{
Extract the data for the 50 US states and the District of Columbia from a dataframe
containing an IHME projection.
}
\examples{
newdf  <- ihme_extractUS(df)
}
\seealso{
\code{ihme_extractWA}
}
