% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ihme_misc.R
\name{ihme_extractWA}
\alias{ihme_extractWA}
\title{Extract the Washington State data from an IHME projection}
\usage{
ihme_extractWA(dat)
}
\arguments{
\item{dat}{A dataframe or tibble containing an IHME projection}
}
\value{
A dataframe containing the IHME projection for Washington State only.
}
\description{
Extract the Washington State data from an IHME projection
}
\examples{
df_wa <- ihme_extractWA(df)
}
