% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ihme_latest.R
\name{ihme_latest}
\alias{ihme_latest}
\title{Download the latest IHME projection}
\usage{
ihme_latest(latestdate = "latest")
}
\arguments{
\item{latestdate}{A character string denoting the date of the projection, e.g. "May 12".
Defaults to "latest" if not specified}
}
\value{
A tibble containing the latest IHME projection
}
\description{
Download the latest IHME projection
}
\examples{
ihme_projection <- ihme_latest()
}
