% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ihme_observed_date.R
\name{ihme_observed_date}
\alias{ihme_observed_date}
\title{Add the date of the last observed actual death counts in an IHME projection time series}
\usage{
ihme_observed_date(dat)
}
\arguments{
\item{dat}{A dataframe or tibble containing an IHME projection.}
}
\value{
A tibble containing an IHME projection, with a column containing the date
of the last observed data in the time series (i.e. the values of
\code{death_means} on that date and earlier dates are the observed death counts;
the values of \code{death_means} on later dates are projected values).
}
\description{
Add the date of the last observed actual death counts in an IHME projection time series
}
\examples{
ihme_observed_date(df)
}
