% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_jhu_data.R
\name{import_jhu_data}
\alias{import_jhu_data}
\title{Import John's Hopkins COVID-19 data}
\usage{
import_jhu_data(type)
}
\arguments{
\item{type}{what dataset do you want returned? cases, recoveries or deaths}
}
\description{
Import John's Hopkins COVID-19 data
}
\examples{
\dontrun{
cases <- import_jhu("cases")
}
}
\keyword{internal}
