% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jhu_misc.R
\name{jhu_extractWA}
\alias{jhu_extractWA}
\title{Extract the Washington State data from the JHUCSSE data}
\usage{
jhu_extractWA(dat)
}
\arguments{
\item{dat}{A dataframe or tibble containing JHUCSSE Covid-19 data}
}
\value{
A dataframe containing the JHUCSSE data for Washington State only.
}
\description{
Extract the Washington State data from the JHUCSSE data
}
\examples{
df_wa <- jhu_extractWA(df)
}
