% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jhu_misc.R
\name{jhu_get_data}
\alias{jhu_get_data}
\title{Download the Covid-19 death counts and case counts for the United States from
the JHUCSSE github site.}
\usage{
jhu_get_data()
}
\value{
A tibble containing the death and case counts for the United States in long
format, combined in one file.
}
\description{
Download the Covid-19 death counts and case counts for the United States from
the JHUCSSE github site.
}
\examples{
df <- jhu_get_data()
}
\seealso{
\code{jhudata_to_long}
}
