% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jhudata_to_long.R
\name{jhudata_to_long}
\alias{jhudata_to_long}
\title{Convert a dataframe containing the Covid-19 data from the Johns Hopkins University
Center for Systems Science and Engineering from the native wide format to a long
format, and add a column with daily death counts.}
\usage{
jhudata_to_long(df)
}
\arguments{
\item{df}{A dataframe or tibble with the JHUCSSE data in wide format.}
}
\value{
A tibble with the JHUCSSE data in long format.
The native wide format has a row for each geographic area, with a column for each
date containing the cumulative death count or case count for that date. The
long format returned by this function has row for each date for each geographic
area, and adds a column for the daily counts.

The function also removes all geographic areas except the 50 US states and the
District of Columbia, and removes unneeded fields.
}
\description{
Convert a dataframe containing the Covid-19 data from the Johns Hopkins University
Center for Systems Science and Engineering from the native wide format to a long
format, and add a column with daily death counts.
}
\examples{
newdf <- jhudata_to_long(df)
}
