% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_url.R
\name{make_url}
\alias{make_url}
\title{Helper function to create html strings needed for loop}
\usage{
make_url(x)
}
\arguments{
\item{x}{countryname that you want to create url for (ends in .csv)}
}
\description{
Helper function to create html strings needed for loop
}
\examples{
\dontrun{
test_url <- function(x = "Zambia")
}
}
