% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_links.R
\docType{data}
\name{owid_vax_url}
\alias{owid_vax_url}
\title{OWID vaccination home}
\format{
An object of class \code{character} of length 1.
}
\usage{
owid_vax_url
}
\description{
link to vaccination data main folder
}
\keyword{datasets}
