% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_links.R
\docType{data}
\name{oxford_url}
\alias{oxford_url}
\title{Oxford COVID-19 Stringency Index}
\format{
An object of class \code{character} of length 1.
}
\usage{
oxford_url
}
\description{
API base
}
\details{
Oxford COVID-19 Government Response Tracker (Stringency Index)
}
\keyword{datasets}
