% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pepfar_iso_map}
\alias{pepfar_iso_map}
\title{PEPFAR COUNTRY ISO CODES}
\format{
A data frame with 71 rows and 2 variables:
\describe{
\item{countryname}{PEPFAR country name}
\item{iso}{3 letter ISO code}
}
}
\source{
\url{https://final.datim.org/}
}
\usage{
pepfar_iso_map
}
\description{
A dataframe containing the PEPFAR country name and its associated ISO Code 3
}
\keyword{datasets}
