% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_google_mobility.R
\name{plot_google_mobility}
\alias{plot_google_mobility}
\title{Plot the Google mobility data for the US, or a US state or counties.}
\usage{
plot_google_mobility(df, level1 = "Washington", level2 = NULL)
}
\arguments{
\item{df}{The R dataframe containing the Google Mobility Community Reports data,
created by the function \verb{download_google_mobility().}}

\item{level1}{The level of geographic region for which data are to be plotted.
This is one of 'US', 'states', or a US state name.}

\item{level2}{Default is NULL. The name of a county for which data are to be plotted.
The state must be named in \code{level1.} If NULL, the data are plotted for the geographic
region named in level1.}
}
\value{
A plot object, with a Google Mobility Community Reports plot for the
US or for a single state or county in the US.
}
\description{
Plot the Google mobility data for the US, or a US state or counties.
}
\examples{
gmcrus <- download_google_mobility("US")
gmplot <- plot_google_mobility(df = gmcrus, level1 = "US", level2 = NULL)
gmplot

gmcrwa <- download_google_mobility("Washington")
gmplot1 <- plot_google_mobility(df = gmcrwa, level1 = "Washington", level2 = NULL)
gmplot1

gmplot2 <- plot_google_mobility(df = gmcrwa, level1 = "Washington", level2 = "Spokane County")
gmplot2
}
