% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_covid_gov_measures.R
\name{pull_covid_gov_measures}
\alias{pull_covid_gov_measures}
\title{Pull HDX COVID Government Measures data}
\usage{
pull_covid_gov_measures(natl_lvl_only = TRUE, pepfar_only = TRUE)
}
\arguments{
\item{natl_lvl_only}{limit data to national level, default = TRUE}

\item{pepfar_only}{limit to just PEPFAR countries, default = TRUE}
}
\value{
dataframe of countries and their types of government interventions
by day
}
\description{
Downloads the UNOCHA HDX COVID Government Measures dataset and loads it into
R. The COVID Government Measures dataset "puts together all the measures
implemented by governments worldwide in response to the Coronavirus
pandemic." For more information, see
https://data.humdata.org/dataset/acaps-covid19-government-measures-dataset.
}
\examples{
\dontrun{
df_gov_measures <- pull_covid_gov_measures()
df_restrictions <- df_gov_measures \%>\% 
  filter(countryname == "Nigeria",
         measure \%in\% c("Domestic travel restrictions",
                        "Partial lockdown", "Full lockdown",
                        "Checkpoints within the country",
                        "Curfews"),
         log_type == "Introduction / extension of measures") \%>\% 
  select(countryname, 
         date = date_implemented, 
         restrict_cat = category, 
         restrict_measure = measure) }
         
}
