% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_jhu_covid.R
\name{pull_jhu_covid}
\alias{pull_jhu_covid}
\title{Pull JHU Covid-19 Data}
\usage{
pull_jhu_covid()
}
\description{
Returns a tidy data frame of JHU COVID-19 data. Returns data frame with colums for
COVID-19 confirmed cases, recoveries and deaths.
}
\details{
Pull a tidy version of JHU COVID-19 data
}
