% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_vax_data.R
\name{pull_vax_data}
\alias{pull_vax_data}
\title{Fetch Our World In Data vaccine data}
\source{
\url{https://github.com/owid/covid-19-data}
}
\usage{
pull_vax_data(cntry)
}
\arguments{
\item{cntry}{country, or list of countries, that appear in the our world in
data page}
}
\value{
Returns data frame of total vaccinations by day
}
\description{
Retrieves COVID-19 vaccination data for a given list of countries
}
\details{
Import OWID vaccine data
}
\examples{
\dontrun{
#only one country
  df_vax_zmb <- get_vax_data("Zambia")
#only PEPFAR countries
  cntry_list <- get_vax_countries()
  df_vax_pepfar <- get_vax_data(cntry_list)
#all countries
  cntry_list <- get_vax_countries(pepfar_only = FALSE)
  df_vax <- get_vax_data(cntry_list)
}

}
\seealso{
Other vax: 
\code{\link{get_vax_countries}()}
}
\concept{vax}
