% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/who_covid_pandemic.R
\name{who_pandemic}
\alias{who_pandemic}
\title{Returns a dataframe of when WHO declared COVID-19 a pandemic}
\usage{
who_pandemic()
}
\description{
Returns a dataframe of when WHO declared COVID-19 a pandemic
}
