% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/first_time_flag.R
\name{first_time_flag}
\alias{first_time_flag}
\title{First time flag}
\usage{
first_time_flag(var, value)
}
\description{
Returns an indicator variable for the first time a condition is met in a dataframe.
USeful for tagging when an event started or ended in a panel dataset.
}
\details{
Flags the first time and event occurs
}
