% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vax_countries.R
\name{get_vax_countries}
\alias{get_vax_countries}
\title{Import Our World in Data Vaccine rates

Pull list of countries for which vax data exist}
\usage{
get_vax_countries(pepfar_only = TRUE)
}
\arguments{
\item{pepfar_only}{pepfar_only limit to just PEPFAR countries, default = TRUE}
}
\value{
Returns an object of the countries being tracked by Our World in Data
}
\description{
Import Our World in Data Vaccine rates
}
\examples{
\dontrun{
#all countries
  cntry_list <- get_vax_countries(pepfar_only = FALSE)
  df_vax <- get_vax_data(cntry_list)
#pepfar only countries
  cntry_list <- get_vax_countries()
  df_vax_pepfar <- get_vax_data(cntry_list)
}
}
\seealso{
Other vax: 
\code{\link{get_vax_data}()}
}
\concept{vax}
