% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_vax_data.R
\name{get_vax_data}
\alias{get_vax_data}
\title{Fetch Our World In Data vaccine data}
\usage{
get_vax_data(cntry)
}
\arguments{
\item{cntry}{country, or list of countries, that appear in the our world in
data page}
}
\value{
Returns data frame of total vaccinations by day
}
\description{
Retrieves COVID-19 vaccination data for a given list of countries
}
\details{
Import OWID vaccine data
}
