% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vax_data.R
\name{get_vax_data}
\alias{get_vax_data}
\title{Fetch Our World In Data vaccine data}
\usage{
get_vax_data(cntry)
}
\arguments{
\item{cntry}{list of countries that appear in the our world in data page (ends in .csv)}
}
\description{
Retrieves COVID-19 vaccination data for a given list of countries
}
\details{
Import OWID vaccine data
}
\examples{
\dontrun{
zmb <- get_vax_data(cntry = "Zambia.csv")
}
}
