% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covid_links.R
\docType{data}
\name{jhuRepo}
\alias{jhuRepo}
\title{JHU main repo}
\format{
An object of class \code{character} of length 1.
}
\usage{
jhuRepo
}
\description{
link to rawdata
}
\details{
John's Hopkins Github Repo Links
}
\keyword{datasets}
