% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_jhu_covid.R
\name{pull_jhu_covid}
\alias{pull_jhu_covid}
\title{Pull JHU COVID-19 Case, Recovery, and Death Data}
\usage{
pull_jhu_covid(pepfar_only = TRUE)
}
\arguments{
\item{pepfar_only}{limit to just PEPFAR countries, default = TRUE}
}
\description{
\%>\% Returns a tidy data frame of JHU COVID-19 data. Returns a tidy
data frame with columns for COVID-19 confirmed cases, recoveries and deaths.
}
\examples{
\dontrun{
covid_data <- pull_jhu_covid() }

}
