% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reshape_jhu_covid.R
\name{reshape_jhu_covid}
\alias{reshape_jhu_covid}
\title{Reshape JHU data}
\usage{
reshape_jhu_covid(df)
}
\arguments{
\item{df}{data frame upon which the reshape is to be conducted}
}
\description{
Pivots JHU COVID-19 data from wide to long format. Creates daily cases. Used in pull_jhu_covid.a
}
\details{
Function to reshape Covid-19 data from Johns Hopkins github repo
}
\keyword{internal}
